backend/requirements.txt
flask
flask-cors
pymysql
flask-jwt-extended
google-cloud-texttospeech
python-dotenv